refactor: 메소드명 변경
